cambios email

+++ database/migrations/2024_10_18_232427_10_18_2024_2.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->bigIncrements("id");
             $table->string("name");
             $table->string("brand");
-            $table->integer("price");
+            $table->double("price");
             $table->integer("code");
             $table->rememberToken();
             $table->timestamps();
